Rewrite cmd_drive_translate to mimic robot behavior
The real robot does not apply the cmd_drive messages directly to the
motor input commands. Instead, there is low-level code that limits the
maximum input command variation sent to the motors. After measuring the
response using an oscilloscope, it takes two seconds to change the input
command from -1 to 1, resulting in a max_update_rate of approximately
one input command unit per second.

This node now mimics this behavior and also enforces a minimum command
frequency. If no messages are received within this frequency, the target
velocity is set to zero.